refactor(roles): dedupe role-color helper into shared lib
roles.tsx and org-role-detail.tsx each carried a byte-identical
ROLE_COLORS array, BUILTIN_ROLE_COLORS map, and getRoleColor/
getRoleDotColor hash-based color picker. Move the single copy to
web/lib/role-color.ts and import it from both call sites.
